Test Resources: * Create UniformTestDataSource - provides uniformly distributed numerical lists * Create LargeTestDataSource - provides large lists with near integer overflow sum * Create TestDataSourceTest - validate TestData from sources * Remove StatisticsTestResourcesTest * Remove most data source methods from StatisticsTestResources
